backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
google-generativeai>=0.8,<1.0
Pillow>=11.0,<12.0
pdf2image>=1.17,<2.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
redis>=5.0,<6.0
